getFrame fix

+++ src/main/scala/com/mikrotasarim/api/command/DeviceController.scala
@@ -295,7 +295,7 @@ class DeviceController(device: DeviceInterface) {
     ))
   }
 
-  def getFrame: Array[Int] = {
+  def getFrame: Array[Byte] = {
     setWiresAndTrigger(Map(
       commandWire -> sFsynOpCode
     ))
